feat: move headers into entities
